// src/api/generated/model/flatRateTypeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { FlatRateTypeResponseStatus } from './flatRateTypeResponseStatus';

export interface FlatRateTypeResponse {
  id?: number;
  name?: string;
  defaultPrice?: number;
  status?: FlatRateTypeResponseStatus;
}