// src/api/generated/zod/flat-rate-types/flat-rate-types.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import * as zod from 'zod';


/**
 * @summary Update flat rate type
 */
export const UpdateFlatRateTypeParams = zod.object({
  "id": zod.int()
})




export const UpdateFlatRateTypeBody = zod.object({
  "name": zod.string().min(1),
  "defaultPrice": zod.number().optional()
})

export const UpdateFlatRateTypeResponse = zod.object({
  "success": zod.boolean().optional(),
  "data": zod.object({
  "id": zod.int().optional(),
  "name": zod.string().optional(),
  "defaultPrice": zod.number().optional(),
  "status": zod.enum(['ACTIVE', 'DISABLED']).optional()
}).optional(),
  "message": zod.string().optional()
})

/**
 * @summary Deactivate a flat rate type
 */
export const DeactivateFlatRateTypeParams = zod.object({
  "id": zod.int()
})

export const DeactivateFlatRateTypeResponse = zod.object({
  "success": zod.boolean().optional(),
  "data": zod.unknown().optional(),
  "message": zod.string().optional()
})

/**
 * Fetches a list of active custom flat rate trip types for dropdown selection.
 * @summary Get all active flat rate types
 */
export const GetActiveFlatRateTypesResponse = zod.object({
  "success": zod.boolean().optional(),
  "data": zod.array(zod.object({
  "id": zod.int().optional(),
  "name": zod.string().optional(),
  "defaultPrice": zod.number().optional(),
  "status": zod.enum(['ACTIVE', 'DISABLED']).optional()
})).optional(),
  "message": zod.string().optional()
})

/**
 * Adds a custom flat rate trip type (e.g. 'Wien -> Graz').
 * @summary Create a new flat rate type
 */



export const CreateFlatRateTypeBody = zod.object({
  "name": zod.string().min(1),
  "defaultPrice": zod.number().optional()
})

export const CreateFlatRateTypeResponse = zod.object({
  "success": zod.boolean().optional(),
  "data": zod.object({
  "id": zod.int().optional(),
  "name": zod.string().optional(),
  "defaultPrice": zod.number().optional(),
  "status": zod.enum(['ACTIVE', 'DISABLED']).optional()
}).optional(),
  "message": zod.string().optional()
})

/**
 * @summary Get all flat rate types including inactive
 */
export const GetAllFlatRateTypesResponse = zod.object({
  "success": zod.boolean().optional(),
  "data": zod.array(zod.object({
  "id": zod.int().optional(),
  "name": zod.string().optional(),
  "defaultPrice": zod.number().optional(),
  "status": zod.enum(['ACTIVE', 'DISABLED']).optional()
})).optional(),
  "message": zod.string().optional()
})